2211: Add S3 wrapper
