+++ stack_tests/integration_tests/docker-compose.yml
@@ -81,7 +81,7 @@ services:
             - POSTGRES_PORT=5435
 
     localstack:
-        image: localstack/localstack
+        image: localstack/localstack:4.14
         hostname: localstack
         ports:
             - "4567:4566"
